Comment out cutout time in config
When the user specifies a snapshot range and tries to download a
cutout for the range, they often doesn't know that the time range
is taken with priority from the "time" in the atlite part of
the config. By default comment this atlite time specification out,
meaning the snapshot range will be used as cutout range.

Comments are added to the config to help the user understanding
some extra atlite functionalities that can be parsed.